+++ common/src/main/java/mca/entity/VillagerEntityMCA.java
@@ -214,10 +214,10 @@ public class VillagerEntityMCA extends VillagerEntity implements VillagerLike<Vi
             if (spawnReason != SpawnReason.BREEDING) {
                 genetics.randomize();
                 traits.randomize();
+            }
 
-                if (spawnReason != SpawnReason.SPAWN_EGG && spawnReason != SpawnReason.DISPENSER) {
-                    genetics.setGender(Gender.getRandom());
-                }
+            if (genetics.getGender() == Gender.UNASSIGNED) {
+                genetics.setGender(Gender.getRandom());
             }
 
             if (Strings.isNullOrEmpty(getTrackedValue(VILLAGER_NAME))) {